Use EntityType in Mob Constructors.

+++ src/MiNET/MiNET/Entities/Hostile/CaveSpider.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class CaveSpider : HostileMob
 	{
-		public CaveSpider(Level level) : base((int) EntityType.CaveSpider, level)
+		public CaveSpider(Level level) : base(EntityType.CaveSpider, level)
 		{
 			Width = Length = 0.7;
 			Height = 0.5;
