+++ src/MiNET/MiNET/Entities/Hostile/Creeper.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Creeper : HostileMob
 	{
-		public Creeper(Level level) : base((int) EntityType.Creeper, level)
+		public Creeper(Level level) : base(EntityType.Creeper, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
