+++ src/MiNET/MiNET/Entities/Hostile/Enderman.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Enderman : HostileMob
 	{
-		public Enderman(Level level) : base((int) EntityType.Enderman, level)
+		public Enderman(Level level) : base(EntityType.Enderman, level)
 		{
 			Width = Length = 0.6;
 			Height = 2.9;
