+++ src/MiNET/MiNET/Entities/Passive/IronGolem.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Passive
 {
 	public class IronGolem : PassiveMob
 	{
-		public IronGolem(Level level) : base((int) EntityType.IronGolem, level)
+		public IronGolem(Level level) : base(EntityType.IronGolem, level)
 		{
 			Width = Length = 1.4;
 			Height = 2.9;
