+++ src/MiNET/MiNET/Entities/Passive/Ocelot.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Passive
 {
 	public class Ocelot : PassiveMob
 	{
-		public Ocelot(Level level) : base((int) EntityType.Ocelot, level)
+		public Ocelot(Level level) : base(EntityType.Ocelot, level)
 		{
 			Width = Length = 0.6;
 			Height = 0.8;
