+++ src/MiNET/MiNET/Entities/Passive/Sheep.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Passive
 {
 	public class Sheep : PassiveMob, IAgeable
 	{
-		public Sheep(Level level) : base((int) EntityType.Sheep, level)
+		public Sheep(Level level) : base(EntityType.Sheep, level)
 		{
 			Width = Length = 0.9;
 			Height = 1.3;
